requirements.txt
Flask 
mysql-connector-python 
